backend/requirements.txt
fastapi
uvicorn
paho-mqtt
python-dotenv
psycopg2-binary
requests
paho-mqtt
